Move some label functions out of utils, move read ai system config into own module
