+++ domino/aisystems/__init__.py
@@ -1,4 +1,4 @@
-from ._util import read_ai_system_config
+from .read_ai_system_config import read_ai_system_config
 
 __all__ = [
     'read_ai_system_config',
